tests/main-4-getQuESTLibraryData.c  fixed typo

+++ tests/main-4-getQuESTLibraryData.c
@@ -10,11 +10,11 @@ int main(int, char*[]) {
        getQuESTLibraryData(quenv, &libdat);
 
        printf("Test correctness of the precision info...\n"); fflush(stdout);
-       if ( libdat.QuEST_Prec == sizeof(qreal)/4 ) {
+       if ( libdat.QuEST_Prec == (int)sizeof(qreal)/4 ) {
             printf("... fine.\n");
        } else {
             printf("`getQuESTLibraryData()` gave %d but truth is %d.\n",
-                   libdat.QuEST_Prec, sizeof(qreal)/4);
+                   (int)libdat.QuEST_Prec, (int)sizeof(qreal)/4);
        }
 
        printf("Test correctness of the rest via `getEnvironmentString()`...\n"); fflush(stdout);
